fixing comp hours balance and request condition

+++ database/migrations/2022_01_08_181141_create_balances_table.php
@@ -18,7 +18,7 @@ class CreateBalancesTable extends Migration
             $table->string('name');
             $table->foreignId('leavetype_id');
             $table->foreignId('user_id');
-            $table->float('value')->nullable();
+            $table->decimal('value',8,3)->nullable();
             $table->timestamps();
         });
     }
